+++ SSRepository/Models/BranchModel.cs
@@ -21,7 +21,7 @@ namespace SSRepository.Models
         [Phone]
         public string? Mobile { get; set; }
         public string? Address { get; set; }
-        public int? FkCityId { get; set; }
+        public long? FkCityId { get; set; }
         public string? State { get; set; }
         public string? Pin { get; set; }
         public string? Country { get; set; }
